// types.ts
export type HolidayDateType = 'specific' | 'recurring' | 'range';

/**
 * Representa una fecha festiva, que puede ser un día específico, un día anual recurrente o un rango de fechas.
 */
export interface HolidayDate {
    type: HolidayDateType;
    day?: number;
    month?: number;
    year?: number;
    startDate?: string; // "YYYY-MM-DD"
    endDate?: string; // "YYYY-MM-DD"
}

/**
 * Representa un día festivo.
 */
export interface Holiday {
  id: number;
  name: string;
  date: HolidayDate;
  location?: string; // Si el festivo es específico de una ubicación
}